requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
pillow
django-environ==0.11.2
psycopg2>=2.9.9,<=2.10
drf-spectacular